separate preview unit tests, relocate sample videos
