+++ README.md
@@ -87,6 +87,7 @@ pysubs2         (for subtitles, comes installed)
 yt_dlp          (for streaming Youtube videos)
 decord          (for videos in bytes, best option)
 imageio         (for videos in bytes)
+pyav            (required for imageio)
 pyglet          (graphics library)
 PySide6         (graphics library)
 PyQt6           (graphics library)
