+++ examples/gif_demo.py
@@ -5,6 +5,9 @@ Gifs are essentially treated as videos with no sound
 Uses looping_demo.py to loop the gif
 '''
 
+# Sample videos can be found here: https://github.com/anrayliu/pyvidplayer2-test-resources/tree/main/resources
+
+
 import pygame
 from pyvidplayer2 import VideoPlayer, Video
