+++ examples/iterate_frames_demo.py
@@ -2,6 +2,9 @@
 This example shows how the Video object can be used as a generator to iterate through each frame
 '''
 
+# Sample videos can be found here: https://github.com/anrayliu/pyvidplayer2-test-resources/tree/main/resources
+
+
 from pyvidplayer2 import Video
