+++ examples/looping_demo.py
@@ -5,6 +5,8 @@ v0.9.26 brought significant improvements to looping
 looping is now much more seamless
 '''
 
+# Sample videos can be found here: https://github.com/anrayliu/pyvidplayer2-test-resources/tree/main/resources
+
 
 import pygame
 from pyvidplayer2 import VideoPlayer, Video
