+++ examples/pip_demo.py
@@ -5,6 +5,8 @@ This is a Picture-in-Picture app
 install pywin32 via pip before using
 '''
 
+# Sample videos can be found here: https://github.com/anrayliu/pyvidplayer2-test-resources/tree/main/resources
+
 
 import pygame
 from win32gui import SetWindowPos, GetCursorPos, GetWindowRect, GetForegroundWindow, SetForegroundWindow
