+++ examples/preview_thumbnails_demo.py
@@ -2,6 +2,8 @@
 This example shows off preview thumbnails in the video player
 '''
 
+# Sample videos can be found here: https://github.com/anrayliu/pyvidplayer2-test-resources/tree/main/resources
+
 
 import pygame
 from pyvidplayer2 import VideoPlayer, Video
