+++ examples/raylib_demo.py
@@ -2,6 +2,9 @@
 This is a quick example of integrating a video into a raylib project
 '''
 
+# Sample videos can be found here: https://github.com/anrayliu/pyvidplayer2-test-resources/tree/main/resources
+
+
 import pyray
 from pyvidplayer2 import VideoRaylib
