+++ examples/seeking_demo.py
@@ -2,6 +2,9 @@
 This example shows the two ways of seeking
 '''
 
+# Sample videos can be found here: https://github.com/anrayliu/pyvidplayer2-test-resources/tree/main/resources
+
+
 from pyvidplayer2 import Video
