+++ examples/subtitles_demo.py
@@ -2,6 +2,8 @@
 This is an example showing how to add subtitles to a video
 '''
 
+# Sample videos can be found here: https://github.com/anrayliu/pyvidplayer2-test-resources/tree/main/resources
+
 
 from pyvidplayer2 import Subtitles, Video
 from pygame import Font
