+++ examples/webcam_app_demo.py
@@ -4,6 +4,9 @@ This is an example showing off a draggable webcam app
 install pywin32 via pip before using
 '''
 
+# Sample videos can be found here: https://github.com/anrayliu/pyvidplayer2-test-resources/tree/main/resources
+
+
 import pygame
 from pyvidplayer2 import Webcam
 import win32api, win32gui, win32con
