+++ setup.py
@@ -20,6 +20,24 @@ setup(
                     "pygame",
                     "pysubs2",
                     "PyAudio"],
+    extras_require={
+        "all": [
+            "numpy",
+            "opencv-python",
+            "pygame",
+            "PyAudio",
+            "pysubs2",
+            "yt-dlp",
+            "decord",
+            "imageio",
+            "pyav",
+            "pyglet",
+            "PySide6",
+            "PyQt6",
+            "raylib",
+            "wxPython"
+        ],
+    },
     url="https://github.com/anrayliu/pyvidplayer2",
     platforms=["windows", "linux", "macos"],
     keywords=["pygame", "video", "playback", "tkinter", "pyqt", "pyside", "pyglet", "raylib", "wxpython", "youtube", "stream"],
